+++ CScape.Core/Network/Handler/IPacketHandler.cs
@@ -5,7 +5,7 @@ namespace CScape.Core.Network.Handler
 {
     public interface IPacketHandler
     {
-        int[] Handles { get; }
+        byte[] Handles { get; }
         void Handle([NotNull] Game.Entity.Player player, int opcode, [NotNull] Blob packet);
     }
 }
